+++ apps/meteor/server/methods/toggleFavorite.ts
@@ -4,6 +4,8 @@ import { Subscriptions } from '@rocket.chat/models';
 import { Match, check } from 'meteor/check';
 import { Meteor } from 'meteor/meteor';
 
+import { notifyOnSubscriptionChangedByRoomIdAndUserId } from '../../app/lib/server/lib/notifyListener';
+
 declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
 	interface ServerMethods {
@@ -28,6 +30,12 @@ Meteor.methods<ServerMethods>({
 			throw new Meteor.Error('error-invalid-subscription', 'You must be part of a room to favorite it', { method: 'toggleFavorite' });
 		}
 
-		return (await Subscriptions.setFavoriteByRoomIdAndUserId(rid, userId, f)).modifiedCount;
+		const { modifiedCount } = await Subscriptions.setFavoriteByRoomIdAndUserId(rid, userId, f);
+
+		if (modifiedCount) {
+			void notifyOnSubscriptionChangedByRoomIdAndUserId(rid, userId);
+		}
+
+		return modifiedCount;
 	},
 });