+++ apps/meteor/server/models/raw/LivechatRooms.ts
@@ -9,7 +9,7 @@ import type {
 	ReportResult,
 	MACStats,
 } from '@rocket.chat/core-typings';
-import { isMessageFromVisitor, UserStatus } from '@rocket.chat/core-typings';
+import { UserStatus } from '@rocket.chat/core-typings';
 import type { ILivechatRoomsModel } from '@rocket.chat/model-typings';
 import type { Updater } from '@rocket.chat/models';
 import { Settings } from '@rocket.chat/models';
@@ -26,6 +26,7 @@ import type {
 	FindCursor,
 	UpdateResult,
 	AggregationCursor,
+	UpdateOptions,
 } from 'mongodb';
 
 import { getValue } from '../../../app/settings/server/raw';
@@ -76,10 +77,6 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 			{ key: { servedBy: 1, ts: 1 }, partialFilterExpression: { servedBy: { $exists: true }, t: 'l' } },
 			{ key: { 'v.activity': 1, 'ts': 1 }, partialFilterExpression: { 'v.activity': { $exists: true }, 't': 'l' } },
 		];
-	}
-
-	getUpdater(): Updater<IOmnichannelRoom> {
-		return super.getUpdater();
 	}
 
 	getQueueMetrics({
@@ -1593,7 +1590,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 		);
 	}
 
-	closeRoomById(roomId: string, closeInfo: IOmnichannelRoomClosingInfo) {
+	closeRoomById(roomId: string, closeInfo: IOmnichannelRoomClosingInfo, options?: UpdateOptions) {
 		const { closer, closedBy, closedAt, chatDuration, serviceTimeDuration, tags } = closeInfo;
 
 		return this.updateOne(
@@ -1615,6 +1612,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 					open: 1,
 				},
 			},
+			options,
 		);
 	}
 
@@ -1977,52 +1975,20 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 		return this.find(query, options);
 	}
 
-	setResponseByRoomId(roomId: string, responseBy: IOmnichannelRoom['responseBy']) {
-		return this.updateOne(
-			{
-				_id: roomId,
-				t: 'l',
-			},
-			{
-				$set: {
-					responseBy,
-				},
-				$unset: {
-					waitingResponse: 1,
-				},
-			},
-		);
+	getResponseByRoomIdUpdateQuery(responseBy: IOmnichannelRoom['responseBy'], updater: Updater<IOmnichannelRoom> = this.getUpdater()) {
+		updater.set('responseBy', responseBy);
+		updater.unset('waitingResponse');
+		return updater;
 	}
 
-	setNotResponseByRoomId(roomId: string) {
-		return this.updateOne(
-			{
-				_id: roomId,
-				t: 'l',
-			},
-			{
-				$set: {
-					waitingResponse: true,
-				},
-				$unset: {
-					responseBy: 1,
-				},
-			},
-		);
+	getNotResponseByRoomIdUpdateQuery(updater: Updater<IOmnichannelRoom> = this.getUpdater()) {
+		updater.set('waitingResponse', true);
+		updater.unset('responseBy');
+		return updater;
 	}
 
-	setAgentLastMessageTs(roomId: string) {
-		return this.updateOne(
-			{
-				_id: roomId,
-				t: 'l',
-			},
-			{
-				$set: {
-					'responseBy.lastMessageTs': new Date(),
-				},
-			},
-		);
+	getAgentLastMessageTsUpdateQuery(updater: Updater<IOmnichannelRoom> = this.getUpdater()) {
+		return updater.set('responseBy.lastMessageTs', new Date());
 	}
 
 	private getAnalyticsUpdateQuery(
@@ -2046,7 +2012,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 		return updater;
 	}
 
-	private getAnalyticsUpdateQueryBySentByAgent(
+	getAnalyticsUpdateQueryBySentByAgent(
 		room: IOmnichannelRoom,
 		message: IMessage,
 		analyticsData: Record<string, string | number | Date> | undefined,
@@ -2063,10 +2029,9 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 		return this.getAnalyticsUpdateQuery(analyticsData, updater);
 	}
 
-	private getAnalyticsUpdateQueryBySentByVisitor(
+	getAnalyticsUpdateQueryBySentByVisitor(
 		room: IOmnichannelRoom,
 		message: IMessage,
-		analyticsData: Record<string, string | number | Date> | undefined,
 		updater: Updater<IOmnichannelRoom> = this.getUpdater(),
 	) {
 		// livechat analytics : update last message timestamps
@@ -2075,29 +2040,18 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 
 		// update visitor timestamp, only if its new inquiry and not continuing message
 		if (agentLastReply >= visitorLastQuery) {
-			return this.getAnalyticsUpdateQuery(analyticsData, updater).set('metrics.v.lq', message.ts);
+			return updater.set('metrics.v.lq', message.ts);
 		}
 
-		return this.getAnalyticsUpdateQuery(analyticsData, updater);
+		return updater;
 	}
 
-	async getAnalyticsUpdateQueryByRoomId(
-		room: IOmnichannelRoom,
-		message: IMessage,
-		analyticsData: Record<string, string | number | Date> | undefined,
-		updater: Updater<IOmnichannelRoom> = this.getUpdater(),
-	) {
-		return isMessageFromVisitor(message)
-			? this.getAnalyticsUpdateQueryBySentByVisitor(room, message, analyticsData, updater)
-			: this.getAnalyticsUpdateQueryBySentByAgent(room, message, analyticsData, updater);
-	}
-
-	getTotalConversationsBetweenDate(t: 'l', date: { gte: Date; lt: Date }, { departmentId }: { departmentId?: string } = {}) {
+	getTotalConversationsBetweenDate(t: 'l', date: { gte: Date; lte: Date }, { departmentId }: { departmentId?: string } = {}) {
 		const query: Filter<IOmnichannelRoom> = {
 			t,
 			ts: {
 				$gte: new Date(date.gte), // ISO Date, ts >= date.gte
-				$lt: new Date(date.lt), // ISODate, ts < date.lt
+				$lte: new Date(date.lte), // ISODate, ts <= date.lte
 			},
 			...(departmentId && departmentId !== 'undefined' && { departmentId }),
 		};
@@ -2107,7 +2061,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 
 	getAnalyticsMetricsBetweenDate(
 		t: 'l',
-		date: { gte: Date; lt: Date },
+		date: { gte: Date; lte: Date },
 		{ departmentId }: { departmentId?: string } = {},
 		extraQuery: Document = {},
 	) {
@@ -2115,7 +2069,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 			t,
 			ts: {
 				$gte: new Date(date.gte), // ISO Date, ts >= date.gte
-				$lt: new Date(date.lt), // ISODate, ts < date.lt
+				$lte: new Date(date.lte), // ISODate, ts <= date.lte
 			},
 			...(departmentId && departmentId !== 'undefined' && { departmentId }),
 			...extraQuery,
@@ -2128,7 +2082,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 
 	getAnalyticsMetricsBetweenDateWithMessages(
 		t: string,
-		date: { gte: Date; lt: Date },
+		date: { gte: Date; lte: Date },
 		{ departmentId }: { departmentId?: string } = {},
 		extraQuery: Document = {},
 		extraMatchers: Document = {},
@@ -2140,7 +2094,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 						t,
 						ts: {
 							$gte: new Date(date.gte), // ISO Date, ts >= date.gte
-							$lt: new Date(date.lt), // ISODate, ts < date.lt
+							$lte: new Date(date.lte), // ISODate, ts <= date.lte
 						},
 						...(departmentId && departmentId !== 'undefined' && { departmentId }),
 						...extraMatchers,
@@ -2210,7 +2164,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 		);
 	}
 
-	getAnalyticsBetweenDate(date: { gte: Date; lt: Date }, { departmentId }: { departmentId?: string } = {}) {
+	getAnalyticsBetweenDate(date: { gte: Date; lte: Date }, { departmentId }: { departmentId?: string } = {}) {
 		return this.col.aggregate<Pick<IOmnichannelRoom, 'ts' | 'departmentId' | 'open' | 'servedBy' | 'metrics' | 'msgs' | 'onHold'>>(
 			[
 				{
@@ -2218,7 +2172,7 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 						t: 'l',
 						ts: {
 							$gte: new Date(date.gte), // ISO Date, ts >= date.gte
-							$lt: new Date(date.lt), // ISODate, ts < date.lt
+							$lte: new Date(date.lte), // ISODate, ts <= date.lte
 						},
 						...(departmentId && departmentId !== 'undefined' && { departmentId }),
 					},
@@ -2456,18 +2410,17 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 		return this.updateOne(query, update);
 	}
 
+	getVisitorActiveForPeriodUpdateQuery(period: string, updater: Updater<IOmnichannelRoom> = this.getUpdater()): Updater<IOmnichannelRoom> {
+		return updater.addToSet('v.activity', period);
+	}
+
 	markVisitorActiveForPeriod(rid: string, period: string): Promise<UpdateResult> {
 		const query = {
 			_id: rid,
 		};
+		const updater = this.getVisitorActiveForPeriodUpdateQuery(period);
 
-		const update = {
-			$addToSet: {
-				'v.activity': period,
-			},
-		};
-
-		return this.updateOne(query, update);
+		return this.updateOne(query, updater.getUpdateFilter());
 	}
 
 	async getMACStatisticsForPeriod(period: string): Promise<MACStats[]> {
@@ -2634,6 +2587,13 @@ export class LivechatRoomsRaw extends BaseRaw<IOmnichannelRoom> implements ILive
 	}
 
 	findOpenRoomsByPriorityId(_priorityId: string): FindCursor<IOmnichannelRoom> {
+		throw new Error('Method not implemented.');
+	}
+
+	getPredictedVisitorAbandonmentByRoomIdUpdateQuery(
+		_willBeAbandonedAt: Date,
+		_updater: Updater<IOmnichannelRoom>,
+	): Updater<IOmnichannelRoom> {
 		throw new Error('Method not implemented.');
 	}